+++ opennmt/utils/misc.py
@@ -7,6 +7,7 @@ import sys
 import inspect
 import six
 
+import numpy as np
 import tensorflow as tf
 
 
@@ -23,6 +24,37 @@ def print_bytes(str_as_bytes, stream=None):
   write_buffer.write(str_as_bytes)
   write_buffer.write(b"\n")
   stream.flush()
+
+def format_translation_output(sentence,
+                              score=None,
+                              token_level_scores=None,
+                              attention=None,
+                              alignment_type=None):
+  """Formats a translation output with possibly scores, alignments, etc., e.g:
+
+  1.123214 ||| Hello world ||| 0.30907777 0.030488174 ||| 0-0 1-1
+
+  Args:
+    sentence: The translation to output.
+    score: If set, attach the score.
+    token_level_scores: If set, attach the token level scores.
+    attention: The attention vector.
+    alignment_type: The type of alignments to format (can be: "hard").
+  """
+  if score is not None:
+    sentence = "%f ||| %s" % (score, sentence)
+  if token_level_scores is not None:
+    scores_str = " ".join(str(f) for f in token_level_scores)
+    sentence = "%s ||| %s" % (sentence, scores_str)
+  if attention is not None and alignment_type is not None:
+    if alignment_type == "hard":
+      source_indices = np.argmax(attention, axis=-1)
+      target_indices = range(attention.shape[0])
+      pairs = ("%d-%d" % (src, tgt) for src, tgt in zip(source_indices, target_indices))
+      sentence = "%s ||| %s" % (sentence, " ".join(pairs))
+    else:
+      raise ValueError("Invalid alignment type %s" % alignment_type)
+  return sentence
 
 def item_or_tuple(x):
   """Returns :obj:`x` as a tuple or its single element."""